+++ async/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `object_store` traits ([#363](https://github.com/stac-utils/stac-rs/pull/363))
+
 ## [0.5.3] - 2024-09-05
 
 ### Changed