+++ async/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.4] - 2024-09-20
+
+⚠️ This crate is now deprecated, use [stac](https://crates.io/crates/stac) instead.
 
 ## [0.5.3] - 2024-09-05
 
@@ -72,7 +74,7 @@ No changes.
 
 Initial release of **stac-async**.
 
-[Unreleased]: https://github.com/stac-utils/stac-rs/compare/stac-async-v0.5.3...main
+[0.5.4]: https://github.com/stac-utils/stac-rs/compare/stac-async-v0.5.3...stac-async-v0.5.4
 [0.5.3]: https://github.com/stac-utils/stac-rs/compare/stac-async-v0.5.2...stac-async-v0.5.3
 [0.5.2]: https://github.com/stac-utils/stac-rs/compare/stac-async-v0.5.1...stac-async-v0.5.2
 [0.5.1]: https://github.com/stac-utils/stac-rs/compare/stac-async-v0.5.0...stac-async-v0.5.1